Minor Changes to remove errors after compiling (#51)
* Fix "missing folder" errors

It doesn't make sense whey this works differently then just
if not DirAccess.dir_exists_absolute(path)
But for some reason this method doesn't skip zips, feel free to test.

* Fix errors that occur after compiling

Creating folders in res:// for humanzier to look through, this prevents a number of errors from occuring. This may be better if we look through the properties and make it dynamic.